requirements.txt
django
tensorflow
Pillow